+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+- **Fix:** Improve error messages when front matter parsing fails.
+
 ## 1.5.1
 
 - Remove webpack-chunk-hash dependency. Recent versions of Webpack do not require this module for deterministic filename hashing.